1. 필터링 2차 완료 - 구동은 가능한 버전 - 디테일 잡아야함

+++ src/main/java/com/onestack/project/domain/Professional.java
@@ -25,7 +25,7 @@ public class Professional {
     private String screeningMsg;
     private Timestamp proDate;
     private String contactableTime;
-    private BigDecimal averagePrice;
+    private int averagePrice;
     private int reviewCount;
 }
